+++ casda_download/RELEASES.md
@@ -2,6 +2,11 @@
 
 All notable changes to the Docker image `docker://aussrc/casda_download` will be documented in this file.
 
+## [v1.0.1]
+
+- Download from point of failure with retries for all files
+- Use keyring on setonix to allow for recent versions of astroquery > 0.4.6
+
 ## [v1.0.0]
 
 - Default queries for downloading relevant files for the following ASKAP projects: EMU, WALLABY, POSSUM and DINGO